refactor: rename files and update endpoints for consistency
- Renamed files for better naming consistency.
- Updated fetchGameSize endpoint to getGameSize.
- Adjusted imports and references.
