+++ src/server/routes/cardRoute.ts
@@ -1,10 +1,10 @@
 import express from 'express'
-import * as cardsController from '../controllers/cards'
+import * as cardsController from '../controllers/cardController'
 
 const router = express.Router()
 
 router.get('/getCards', cardsController.getCards)
-router.get('/fetchGameSize', cardsController.fetchGameSize)
+router.get('/getGameSize', cardsController.getGameSize)
 router.post('/addCard', cardsController.addCard)
 router.put('/updateCard/:id', cardsController.updateCard)
 router.delete('/deleteCard/:id', cardsController.deleteCard)
